requirements.txt
django==3.2.12
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
py-yaml-fixtures==0.6.1
pylint